Fix mismatched-tags and sign-compare compiler warnings
- Use struct instead of class for SDL_Renderer forward declaration in cInfoContextCreator.h
- Use struct instead of class for sGameServices forward declaration in cGame.h
- Use size_t instead of int for loop counter comparing against bulletInfos.size()

+++ src/context/cInfoContextCreator.cpp
@@ -856,7 +856,7 @@ void cInfoContextCreator::initBullets(cBulletInfos& bulletInfos)
 {
     logbook("Installing:  BULLET TYPES");
 
-    for (int i = 0; i < bulletInfos.size(); i++) {
+    for (size_t i = 0; i < bulletInfos.size(); i++) {
         bulletInfos[i].bmp = nullptr; // in case an invalid bitmap; default is a small rocket
         bulletInfos[i].moveSpeed = 2;
         bulletInfos[i].deathParticle = -1; // this points to a bitmap (in data file, using index)
